Xpad-0.1.4 port from 2.4 to 2.6.
Currently only the joystick and mouse devices supported.
DVD dongle will be a separate module again later.

I noticed a possible problem with the mouse support. If you happen to
encounter that the buttons work, while moving the pointer doesn't, try to
reload the module and tell me about the problem.