ChangeSet@1.1938.452.13  2004-12-20 15:43:59-08:00  greg@kroah.com
USB: convert the idVendor, idProduct, bcdDevice and bcdUSB fields to __le16

These fields are in the struct usb_device_descriptor, and now we keep the
native (on-the-wire mode) format of these fields.  Any driver using these
fields needs to convert it to cpu endian before using them.

All USB drivers in the kernel tree have been fixed up to work properly with
this change.  All out-of-the USB kernel drivers are on their own...

Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>